server/pyproject.toml
[tool.poetry]
name = "cloud-raiders"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.85.0"
uvicorn = "0.29.0"
python-jose = "^3.3.0"
sqlmodel = "^0.0.16"
gunicorn = "^22.0.0"
python-dotenv = "^1.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.ruff]
line-length = 97
indent-width = 4
exclude = [".git", "__pycache__", ".venv", ".vscode"]
target-version = "py312"

[tool.ruff.lint]
select = ["C", "E", "F", "W", "B"]
ignore = ["E501", "E203", "B008"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401"] # import violations

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
